move principia methods to principiawrapper also fix https://github.com/Ezriilc/HyperEdit/issues/73 (i tried adding the principia methods to the landing methods but it kept getting called like 30 times so idk)
